Add operator metadata
