+++ src/AllenNeuralDynamics.Core.Design/PushButtonSource.cs
@@ -3,9 +3,9 @@ using System.ComponentModel;
 
 namespace AllenNeuralDynamics.Core.Design
 {
-    [TypeVisualizer(typeof(AnnotationSourceVisualizer))]
+    [TypeVisualizer(typeof(PushButtonVisualizer))]
     [Description("Generates a sequence of manual annotations or alerts about the experiment.")]
-    public class AnnotationSource : Annotation<string>
+    public class PushButtonSource : PushButton
     {
     }
 }
